Add initial files
These are the files I have been working on. I have implemented all of
the functions, but there is still work I want to do: I want to make them
ANSI C compatible, I want there to be lots of comments explaining the
code, and I want the code to be in the same code style as ClassiCube, so
I can add them into that code base (they want to get rid of libm
dependence).